Result sets larger than 2 GB threw an exception "Negative seek offset".

+++ h2/src/main/org/h2/log/UndoLogRecord.java
@@ -149,7 +149,7 @@ public class UndoLogRecord {
      * @param file the file
      */
     void seek(FileStore file) throws SQLException {
-        file.seek(filePos * Constants.FILE_BLOCK_SIZE);
+        file.seek(((long) filePos) * Constants.FILE_BLOCK_SIZE);
     }
 
     /**
